+++ CHANGELOG.md
@@ -48,6 +48,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Downloads of embeded videos and direct URLs (Imagepond)
 - Thumbnails downloads always failing with 404 (Cyberdrop)
 - Downloads failing with "DDoS-Guard" errors (Yandex Disk)
+- Corrupted file when resuming a partial download that was already 50% or more done
 
 ## [10.8.0] - 2026-09-06